requirements.txt
numpy>=1.7.1
scipy
h5py>=2.0.1
httplib2>=0.7.0
networkx>=1.7
matplotlib
simplekml
simplejson
geomag>=0.9
PyYAML
pytz
FlightDataUtilities
HDFAccess
python-dateutil